Auto-detect series in metadata support
In preparation for the switch to series in metadata across
all OpenStack charms, update the charm store pusher to auto
detect series in metadata, and publish charms using the multi
target, allowing the charm store to figure out which series
a charm should support.

If detected, this is preferred over any configuration files
in release-tools.