Add finding duplicate reviews
I did a bad thing and managed to create lots of duplicate reviews.  I
wrote the scripts in this patch to find them so I could deal with them.
They will probably be useful in the future.

(I'm beginning to feel that we should re-write release tools into a
single python application).