Pin charm-tools (for build) to 2.8.3 for 21.04 cycle
This PR pins charm-tools to 2.8.3 for the reactive builds for the 21.04
development cycle.  The tox.ini is modified to keep the builds in the
build/builds/<charm>  directory as this has been 'fixed' in 2.8.2+ to
build/<charm>.  In order to keep the release tooling consistent between
the stable and master branches the build/builds/<charm> option is used.