Wrap pip to force pip<20.3 and setuptools<50.0.0
Osci runs the CI jobs on Bionic, this version of tox (2.5) will
automatically upgrade the packages at virtualenv creation and any
pinning won't have effect since the virtualenv created is upgraded to
the latest versions of pip and setuptools. This behavior was changed in
tox 3.10 with the addition of the 'download' directive[0]

Setuptools 58.0 dropped the support for use_2to3=true which is needed to
install blessings (an indirect dependency of charm-tools) since then
the gate got broken.

Closes #155

[0] https://github.com/tox-dev/tox/commit/18147f4c5af75caf40c3c2f95a0907da2db83deb