Add commentary and comments for charm/channel searching
In order to avoid reading the yaml file as a yaml file, making changes,
and then writing out the yaml file, the algorithm treats the yaml file
as a series of indented text lines.  The uses a 'trick' that dictionary
keys in the same dictionary are all at the same indent 'depth' in a yaml
file.

Thus, the add_channel_to() function can make minimal changes to the
bundle files, thus making minimal, and easy to (human) check.