Fix push-and-release and add stable -> candidate copier
This fixes push-and-release so that it does a charm grant for the
channel, as apparently the charmstore has to have the channel version
granted.

The 2nd part is a helper script to copy the current stable channel charm
to the candidate release channel (and grant it).  This is to provide a
'starting' point so that charm pull <charm> --channel=candidate has
something to work with.  It would appear that the charmstore randomly
will not provide a charm if a channel is provided that hasn't been
pushed to (which kind of means 'auto-promotion' to a more stable version
doesn't work.)