Add remove-channel feature to script
Add feature to enable removing the channel specification from the charm
sections of a bundle.  This is to enable the bundles to target the
default (stable) charms when a charm is promoted.  This is for release
day activites.

Also, add argument parsing, help and choices to make the script more
robust.