Add in subcommands for showing / working with recipes
This patch starts the process of adding in subcommands to enable showing
and listing the relevant charms and their config to all a use to see
what is happening first.  Only the 'sync' command is destructive and it
requires a i-really-mean-it flag to make the sync/changes happen.
'diff' is 'sync' without the 'sync.