Add 'diff' command
This lets a user see what the difference between the current config in
Launchpad and the desired config in the yaml files.  This can print a
comprehensive diff to show what would need to be done.

Also includes a refactor of the diff code so that it can be shared
between printing and updating/creating charm recipes.