Update update-channel-single.py to work with lp-config
The lp-builder-config yaml files provide git branch -> charmhub
track/channel mappings for the builder config.  This patch updates the
update-channel-single.py tool with an option to provide branches instead
of channels, and then the appropriate charmhub track/channel will be
used for the charm in the bundle.  This is to automate handling of the
charms tests bundle updates.