Updates to global files for reactive charms, charmcraft and charmhub
This patch to the global/source-zaza files is to enable a reactive charm
to be built by the CI system in the 'charm-build' job. This is so that
the charm will have to "charmcraft build" before doing tests and ensures
that it is at least buildable prior to being delivered to launchpad for
build and push to the charmhub.