These are useful tools for doing the charmhub migration
This branch/PR is (a/the) collection of tools to help automate the
charmhub migration for stable/21.10 to multiple branches.  There's lots
of repetition, and as such this is a throwaway branch, where, once
charmhub migration is done, it can be reaped for useful tools for the
next Openstack/Ceph/LTS cycle, but otherwise discardec.

This branch will live as a draft PR, so that it can be used by other
people.