Move openstack-loadbalancer to the misc section
The openstack-loadbalancer isn't actually an openstack service, but is
instead a loadbalancer for openstack projects.  As such, it's more like
rabbitmq-server or mysql-innodb-cluster and will have 'focal' and
'jammy' releases, that are aligned to the LTS releases that it runs on.

This PR moves them.  Note, this will affect upstream branches and tracks
on charmhub.