Changes to lp-builder-config for 22.04 release
This is mainly moving channels to the stable/yoga, stable/22.04
and stable/quincy tracks for openstack, ovn and ceph charms.  The
misc charms are handled a little differently, in that charms need to
span two LTS (for series upgrade), but also try to track the major
version number of the payload.