Modify openstack-loadbalancer to have a stable/jammy channel
Following feedback and some discussion, it's been decided to provide
a 'stable/jammy' branch in the git repo to provide the stable charm for
the 'jammy/edge' channel in charmhub.  Thus master can be used for more
breaking changes.  We may decide to just 'move' the stable/jammy branch
forward for each development release, but that's a discussion for
another time.