Switch recent stable releases to <track>/stable
This is inline with the current policy where a stable backport receives
two +2s on Gerrit and then is immediately pushed to the stable branch.
The stable branches are:

 - yoga/stable
 - quincy/stable
 - 22.03/stable
 - various for the misc charms (see patch)

Also, added an 'enabled: False' key for branches that have not yet been
enabled so that the tool can ignore them unless they are specifically
mentioned.