Stop attempt to pin pip, setuptools from tox
Back in the day, when the new pip resolver came about, the
workaround being removed by this patch was required to move
forward.

Since then the pins have no meaning since newer versions of Python
force the pins to be ignored.

Removing this pinning also allows us to drop the `pip.sh` hack.