Misc/ceph lp-builder-config updates for 22.10
These updates correspond to updates made in this topic:
https://review.opendev.org/q/topic:kinetic-support

I did not update ceph-osd because it already built successfully
and is published to latest/edge for 22.10.

Note that ceph-fs adds all of the stable branches from the default
section as well. As pointed out by Alex, configuration of branches
here overwrites all of the default set of branches.