Treat fetch of master or main as an alias
We have a mix of charm repositories with master and main
branches. Therefore 'fetch-charms.py --branch master' will fail
for some git repos and 'fetch-charms.py --branch main' will do
the same.

This change updates fetch-charms.py to treat master and main as an
alias. If fetching master fails for a project, try to fetch main.
And if fetching main fails for a project, try to fetch master.